Parsing is OK, managing flags are OK, next is norminette (function lengths, modularity, cleaning etc.)

+++ libft/ft_atoi.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aluslu <aluslu@student.42lyon.fr>          +#+  +:+       +#+        */
+/*   By: aluslu <aluslu@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/10 11:39:13 by aluslu            #+#    #+#             */
-/*   Updated: 2025/11/10 15:49:48 by aluslu           ###   ########lyon.fr   */
+/*   Updated: 2025/12/12 16:02:38 by aluslu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,9 +17,9 @@ static int	ft_isspace(char c)
 	return (0);
 }
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	result;
+	long	result;
 	int	sign;
 	int	position;
